backend/requirements.txt
# =============================================================================
# dbt-contracts backend dependencies
# =============================================================================

# --- Web framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0

# --- Async HTTP / SSE ---
httpx==0.27.0
# sse-starlette is NOT used — we use FastAPI's built-in StreamingResponse
# with media_type="text/event-stream" for correct SSE behaviour.
# sse-starlette==2.1.0

# --- Config & validation ---
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# --- Anthropic SDK ---
# Must be >= 0.27 for AsyncAnthropic.messages.stream() with claude-sonnet-4-20250514
# and correct granular exception types (AuthenticationError, RateLimitError, etc.)
anthropic==0.40.0

# --- Supabase ---
supabase==2.4.2

# --- Rate limiting ---
slowapi==0.1.9
limits==3.10.1

# --- YAML parsing (contract validation) ---
PyYAML==6.0.1
yamllint==1.35.1

# --- Utilities ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
orjson==3.10.3

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# --- Dev utilities ---
python-dotenv==1.0.1
